requirements.txt
rick-db>=2.0.2
rick>=0.8.1
rick-mailer>=1.1.1
Flask>=3.1.3
Werkzeug
Flask-Login
iso8601>=2.1.0
setuptools>=75.6.0
tabulate>=0.9.0
pyhumps>=3.8.0
redis>=5.2.1
